DEDA_Class_2019WS_BTC_Blockchain_Analytics/requirements.txt
alabaster==0.7.12
asn1crypto==1.3.0
astroid==2.3.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cloudpickle==1.2.2
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-bigquery==1.23.1
google-cloud-core==1.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
keyring==20.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lightgbm==2.3.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
numpy==1.18.1
numpydoc==0.9.2
packaging==20.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
prompt-toolkit==3.0.2
protobuf==3.11.2
psutil==5.6.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
rope==0.14.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
six==1.13.0
snowballstemmer==2.0.0
soupsieve==1.9.5
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.6
spyder-kernels==0.5.2
stem==1.8.0
testpath==0.4.4
threaded==4.0.8
tornado==6.0.3
torrequest==0.1.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xgboost==0.90
zipp==0.6.0
